Poniendo todo en orden
